Fixes regression caused by FELIX-1488
If one of the three calls inside of the `delete()` method throws an exception, the `isDeleted` field remains true, which causes some unexpected `IllegalStateException`s. 

Before the field was introduced, whenever the `isDeleted()` method was called it would invoke `getPersistenceManager()`, which would then use the local persistence manager to actually check to see if the configuration had been removed from the persistence.

Now the boolean field is being set before any persistence calls actually happen, which is a problem if any of the calls throw.